agent/requirements.txt
psutil
requests
python-socketio>=5.0.0
websocket-client